refactor: Jwt Adapter Divide
