+++ koala-infrastructure/src/main/kotlin/comit/simsimpulyi/koala/global/security/token/JwtAdapter.kt
@@ -3,60 +3,33 @@ package comit.simsimpulyi.koala.global.security.token
 import comit.simsimpulyi.koala.domain.user.model.Authority
 import comit.simsimpulyi.koala.domain.user.spi.ReceiveTokenPort
 import comit.simsimpulyi.koala.global.security.SecurityProperties
-import comit.simsimpulyi.koala.global.security.principle.AuthDetailsService
-import io.jsonwebtoken.*
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken
-import org.springframework.security.core.Authentication
+import io.jsonwebtoken.Header
+import io.jsonwebtoken.Jwts
+import io.jsonwebtoken.SignatureAlgorithm
 import org.springframework.stereotype.Component
 import java.util.*
-import javax.servlet.http.HttpServletRequest
 
 @Component
 class JwtAdapter(
-    private val securityProperties: SecurityProperties,
-    private val authDetailsService: AuthDetailsService
+    private val securityProperties: SecurityProperties
 ) : ReceiveTokenPort {
 
-    override fun generateAccessToken(userId: UUID, authority: Authority): String {
-        return generateToken(userId, authority, securityProperties.accessExpire, "access")
+    override fun generateAccessToken(email: String, authority: Authority): String {
+        return generateToken(email, authority, securityProperties.accessExpire, "access")
     }
 
-    override fun generateRefreshToken(userId: UUID, authority: Authority): String {
-        return generateToken(userId, authority, securityProperties.refreshExpire, "refresh")
+    override fun generateRefreshToken(email: String, authority: Authority): String {
+        return generateToken(email, authority, securityProperties.refreshExpire, "refresh")
     }
 
-    fun generateToken(userId: UUID, authority: Authority, exp: Int, type: String): String {
+    private fun generateToken(email: String, authority: Authority, exp: Int, type: String): String {
         return Jwts.builder()
-            .signWith(SignatureAlgorithm.HS512, securityProperties.secret)
+            .signWith(SignatureAlgorithm.HS512, securityProperties.secretKey)
             .setHeaderParam(Header.JWT_TYPE, type)
-            .setSubject(userId.toString())
+            .setSubject(email)
             .claim("authority", authority)
             .setIssuedAt(Date())
             .setExpiration(Date(System.currentTimeMillis() + exp))
             .compact()
-    }
-
-    fun resolveToken(request: HttpServletRequest): String? {
-        val bearer = request.getHeader("Authentication")
-        if(bearer.isNotEmpty().and(bearer.startsWith("bearer "))) {
-            return bearer.substring(7)
-        }
-        return null
-    }
-
-    fun getClaims(token: String): Jws<Claims> {
-        return Jwts.parser()
-            .setSigningKey(securityProperties.secret)
-            .parseClaimsJws(token)
-    }
-
-    fun getAuthentication(token: String) : Authentication {
-        val claims = getClaims(token)
-
-        if(claims.header[Header.JWT_TYPE] != "access") throw Exception() // TODO
-
-        val detail = authDetailsService.loadUserByUsername(claims.body.subject)
-
-        return UsernamePasswordAuthenticationToken(detail, "", detail.authorities)
     }
 }
